refactor: 정적 파일 구분을 위해 폴더 생성 후 이미지 경로 수정
